Make Elevator and Intake VerticalExtenders
